fix(backend_epoll): build on current V master — cast epoll flags individually
Current V master regressed casting a COMBINED C-constant bitmask to u32 inside
the lib: `u32(C.EPOLLIN | C.EPOLLET)` now errors with "mismatched types u32 and
int / expression does not return a value so it cannot be cast" (it builds in
isolation but not in this module context), which broke every epoll example.
Cast each flag separately — `(u32(C.EPOLLIN) | u32(C.EPOLLET))` — which is
numerically identical and builds on BOTH master and the 0.5.1 release.

Also revert the CI V-pin experiments: the repo requires master V (the database
and hexagonal examples use master's db.pg/ORM APIs and don't build on 0.5.1; the
0.5.1 *tag* won't source-build via setup-v either — `unknown ident native`), so
the workflows stay on setup-v master. With the cast fix all eight matrix examples
(incl. the new async_pipe) build + test on master, and `v fmt -verify .` is clean.

Verified on V master (/usr/local/bin/v, 065a450): all matrix examples build, the
async_pipe e2e test and server_test pass, repo is fmt-clean.

+++ http_server/backend_epoll/conn_state_linux.c.v
@@ -391,7 +391,7 @@ fn park_write(epoll_fd int, fd int, limits core.Limits, mut st PlainState, mut c
 		cs.write_deadline = time.sys_mono_now() + u64(limits.write_timeout_ms) * 1_000_000
 		st.parked++
 	}
-	epoll.mod_fd_in_epoll(epoll_fd, fd, u32(C.EPOLLIN | C.EPOLLOUT | C.EPOLLET))
+	epoll.mod_fd_in_epoll(epoll_fd, fd, (u32(C.EPOLLIN) | u32(C.EPOLLOUT) | u32(C.EPOLLET)))
 }
 
 // append_file_region reads [off, off+len) from a borrowed file fd into `buf`.
@@ -509,7 +509,7 @@ fn handle_writable_plain(epoll_fd int, fd int, active_conns &core.Counter, mut s
 	}
 	if cs.write_off >= cs.write_buf.len && cs.file_remaining <= 0 {
 		// Spurious wake — nothing parked; stop watching writability.
-		epoll.mod_fd_in_epoll(epoll_fd, fd, u32(C.EPOLLIN | C.EPOLLET))
+		epoll.mod_fd_in_epoll(epoll_fd, fd, (u32(C.EPOLLIN) | u32(C.EPOLLET)))
 		return true
 	}
 	// Phase 1: finish the buffered bytes.
@@ -546,7 +546,7 @@ fn handle_writable_plain(epoll_fd int, fd int, active_conns &core.Counter, mut s
 		cs.write_deadline = 0
 		st.parked--
 	}
-	epoll.mod_fd_in_epoll(epoll_fd, fd, u32(C.EPOLLIN | C.EPOLLET)) // stop watching writability
+	epoll.mod_fd_in_epoll(epoll_fd, fd, (u32(C.EPOLLIN) | u32(C.EPOLLET))) // stop watching writability
 	return true
 }
